Har fortsat abstraheringen. Oversat nogle flere dele til Esperanto, og har gjort så tekstboksen virker bedre med fremmedsprog

+++ Chatrum/AddServerPrompt.cs
@@ -8,7 +8,7 @@ namespace Chatrum
     public partial class AddServerPrompt : Form
     {
         public string IP;
-        public int Port;
+        public short Port;
         public string ServerNickname;
 
         public AddServerPrompt()
@@ -34,7 +34,7 @@ namespace Chatrum
 
         private void AddServerBtn_Click(object sender, EventArgs e)
         {
-            labelInvalidPort.Visible = !int.TryParse(ServerPortInput.Text, out Port);
+            labelInvalidPort.Visible = !short.TryParse(ServerPortInput.Text, out Port);
             labelInvalidIP.Visible = !IPAddress.TryParse(ServerIPInput.Text, out _);
 
             if (labelInvalidPort.Visible || labelInvalidIP.Visible || labelInvalidServerNickname.Visible)
